refactor(mobile): split SessionReplay.tsx 617->226 LOC (Cluster A2)
WHY: SessionReplay packed a requestAnimationFrame playback engine, the timing
math, a free-tier upgrade gate, and a speed dropdown into 617 lines - over the
400-LOC ceiling, with the playback-correctness math trapped in useMemo bodies
that can't run in jsdom.

Extracted into a co-located session-replay/ directory:
- replay-timing.ts   - pure timing math (formatTime, toChatMessageData,
                       computeTiming, computeVisibleIndex); now directly testable
- useReplayPlayback.ts- the rAF tick loop + transport state + control handlers,
                       moved verbatim with every ref preserved
- types.ts           - ReplayMessageData, SessionReplayProps, Playback* unions
- constants.ts       - SPEED_OPTIONS, SCREEN_WIDTH
- UpgradePrompt.tsx  - free-tier gate (Apple Reader-App WHY comment preserved)
- SpeedSelector.tsx  - speed dropdown
- __tests__/replay-timing.test.ts - 9 tests pinning duration/offset/visible-index

The orchestrator (226 LOC) re-exports ReplayMessageData so the direct import in
app/session/[id].tsx is unchanged. Dropped a never-rendered visibleMessages
slice (was prefixed _ "for future use"); documented WHY it is safe in the hook.

Verification: mobile typecheck + lint clean; full suite 1747 pass (+9 new);
a11y regression guard green (accessibility props preserved on all touchables).
